Add handleKeyDown and handleKeyUp functions to block navigation logic
